filtration, pagination, sorting

+++ Service/CitizenFormService.cs
@@ -16,12 +16,17 @@ namespace MigrationApi.Service
             _currentUserService = currentUserService;
         }
 
-        public async Task<List<CitizenFormDto>> GetAllAsync()
+
+        public async Task<List<CitizenFormDto>> GetAllAsync(CitizenFormFilterDto filter)
         {
-            var forms = await _repo.GetAllAsync();
-            return forms.Select(f => new CitizenFormDto
+
+
+            var forms = await _repo.GetAllAsync(filter);
+            
+            return forms
+            .Select(f => new CitizenFormDto
             {
-                Id=f.Id,
+                Id = f.Id,
                 RegistrationDate = f.RegistrationDate,
                 PIN = f.PIN,
                 FullName = f.FullName,
@@ -45,6 +50,72 @@ namespace MigrationApi.Service
                     }).ToList()
             }).ToList();
         }
+
+
+        
+        public async Task<List<CitizenFormDto>> GetAllActiveAsync()
+        {
+
+
+            var forms = await _repo.GetAllActiveAsync();
+
+            return forms.Select(f => new CitizenFormDto
+            {
+                Id = f.Id,
+                RegistrationDate = f.RegistrationDate,
+                PIN = f.PIN,
+                FullName = f.FullName,
+                BirthDate = f.BirthDate,
+                Gender = f.Gender,
+                District = f.District == null ? null : new DistrictDto
+                {
+                    Name = f.District.Name,
+                    Region = f.District.Region == null ? null : new CreateRegionDto
+                    {
+                        Name = f.District.Region.Name
+                    }
+                },
+                Migrations = f.Migrations
+                    .OrderByDescending(m => m.DepartureDate)
+                    .Take(1)
+                    .Select(m => new MigrationDto
+                    {
+                        CountryName = m.Country.Name,
+                        DepartureDate = m.DepartureDate
+                    }).ToList()
+            }).ToList();
+        }
+
+        public async Task<List<CitizenFormDto>> GetAllArchivedAsync()
+        {
+            var forms = await _repo.GetAllArchivedAsync();
+            return forms.Select(f => new CitizenFormDto
+            {
+                Id = f.Id,
+                RegistrationDate = f.RegistrationDate,
+                PIN = f.PIN,
+                FullName = f.FullName,
+                BirthDate = f.BirthDate,
+                Gender = f.Gender,
+                District = f.District == null ? null : new DistrictDto
+                {
+                    Name = f.District.Name,
+                    Region = f.District.Region == null ? null : new CreateRegionDto
+                    {
+                        Name = f.District.Region.Name
+                    }
+                },
+                Migrations = f.Migrations
+                    .OrderByDescending(m => m.DepartureDate)
+                    .Take(1)
+                    .Select(m => new MigrationDto
+                    {
+                        CountryName = m.Country.Name,
+                        DepartureDate = m.DepartureDate
+                    }).ToList()
+            }).ToList();
+        }
+
 
         public async Task<OneCitizenFormDto?> GetByIdAsync(Guid id)
         {
@@ -99,11 +170,74 @@ namespace MigrationApi.Service
                 }).ToList(),
                 CreatedByUserName = f.CreatedByUser?.UserName,
                 UpdatedByUserName = f.UpdatedByUser?.UserName,  // <-- имя обновившего
-                CreatedAt = f.Created_at, 
+                CreatedAt = f.Created_at,
                 UpdatedAt = f.Updated_at
-    
+
             };
         }
+        
+
+        public async Task<OneCitizenFormDto?> GetByPINAsync(string PIN)
+        {
+            var f = await _repo.GetByPINAsync(PIN);
+            if (f == null) return null;
+
+            return new OneCitizenFormDto
+            {
+                IsArchived = f.IsArchived,
+                RegistrationDate = f.RegistrationDate,
+                Id = f.Id,
+                ESYBMId = f.ESYBMId,
+                ENIcode = f.ENIcode,
+                PIN = f.PIN,
+                FirstName = f.FirstName,
+                LastName = f.LastName,
+                MidName = f.MidName,
+                BirthDate = f.BirthDate,
+                Gender = f.Gender,
+                Disabilities = f.Disabilities,
+                PhoneNumber = f.PhoneNumber,
+                RuralArea = f.RuralArea,
+                PopulationArea = f.PopulationArea,
+                Street = f.Street,
+                House = f.House,
+                Apartments = f.Apartments,
+
+                District = f.District == null ? null : new DistrictDto
+                {
+                    Name = f.District.Name,
+                    Region = f.District.Region == null ? null : new CreateRegionDto
+                    {
+                        Name = f.District.Region.Name
+                    }
+                },
+                MaritalStatus = f.MaritalStatus == null ? null : new MaritalStatusDto
+                {
+                    Name = f.MaritalStatus.Name
+                },
+                Status = f.Status == null ? null : new StatusDto
+                {
+                    Name = f.Status.Name
+                },
+
+                Migrations = f.Migrations.Select(m => new MigrationHistoryDto
+                {
+                    DepartureDate = m.DepartureDate,
+                    ReturnDate = m.ReturnDate,
+                    Profession = m.Profession,
+                    EmploymentContract = m.EmploymentContract,
+                    CountryName = m.Country.Name
+                }).ToList(),
+                CreatedByUserName = f.CreatedByUser?.UserName,
+                UpdatedByUserName = f.UpdatedByUser?.UserName,  // <-- имя обновившего
+                CreatedAt = f.Created_at,
+                UpdatedAt = f.Updated_at
+
+            };
+        }
+
+
+
 
         public async Task<CitizenForm> CreateAsync(CreateCitizenFormDto dto)
         {
@@ -113,8 +247,8 @@ namespace MigrationApi.Service
             if (!await _repo.MaritalStatusExistsAsync(dto.MaritalStatusId))
                 throw new ArgumentException("Invalid Marital Status Id");
 
-                var userId = _currentUserService.UserId ?? throw new Exception("User not authenticated");
-                
+            var userId = _currentUserService.UserId ?? throw new Exception("User not authenticated");
+
 
             var newform = new CitizenForm
             {
@@ -138,7 +272,7 @@ namespace MigrationApi.Service
                 StatusID = dto.StatusId,
                 Created_at = DateTime.UtcNow,
                 CreatedByUserID = userId
-                
+
             };
 
             await _repo.AddAsync(newform);
@@ -162,6 +296,10 @@ namespace MigrationApi.Service
             if (!await _repo.MaritalStatusExistsAsync(dto.MaritalStatusId))
                 throw new ArgumentException("Invalid Marital Status Id");
 
+            if (form.IsArchived)
+                throw new InvalidOperationException("Archived records cannot be modified.");
+
+
             form.BirthDate = dto.BirthDate;
             form.Gender = dto.Gender;
             form.FirstName = dto.FirstName;
@@ -179,10 +317,10 @@ namespace MigrationApi.Service
             form.StatusID = dto.StatusId;
             form.UpdatedByUserID = userId;
             form.Updated_at = DateTime.UtcNow;
-            
+
 
             form.IsArchived = dto.IsArchived;
-            
+
             await _repo.UpdateAsync(form);
             await _repo.SaveAsync();
             return true;
@@ -193,10 +331,41 @@ namespace MigrationApi.Service
             var form = await _repo.GetByIdAsync(id);
             if (form == null) return false;
 
+            if (form.IsArchived)
+                throw new InvalidOperationException("Archived records cannot be modified.");
+
+
             await _repo.DeleteAsync(form);
+            await _repo.SaveAsync();
+            return true;
+        }
+
+        public async Task<bool> UnarchiveAsync(Guid id)
+        {
+            var citizen = await _repo.GetByIdAsync(id);
+            if (citizen == null) return false;
+
+            if (!citizen.IsArchived)
+                throw new InvalidOperationException("Citizen form is already active.");
+
+            citizen.IsArchived = false;
+            await _repo.UpdateAsync(citizen);
+            await _repo.SaveAsync();
+            return true;
+        }
+        
+        public async Task<bool> ArchiveAsync(Guid id)
+        {
+            var citizen = await _repo.GetByIdAsync(id);
+            if (citizen == null) return false;
+
+            if (citizen.IsArchived)
+                throw new InvalidOperationException("Citizen form is already unactive.");
+
+            citizen.IsArchived = true;
+            await _repo.UpdateAsync(citizen);
             await _repo.SaveAsync();
             return true;
         }
     }
 }
-
